requirements.txt
easyocr
google-generativeai
opencv-python
numpy
paddleocr
paddlepaddle
pillow
pyenchant
shapely

